+++ CHANGELOG.md
@@ -5,16 +5,22 @@ renamed to `LatLng::try_from_dec` to fix name collision with the
 [TryFrom](https://doc.rust-lang.org/std/convert/trait.TryFrom.html) trait.
 Added `try_from_f32` and `try_from_f64` methods for the `LatLng` type.
 
+* 2.2.0: 2022-09-03: Optional basic support for the
+[geo](https://crates.io/crates/geo) crate and [GeoRust](https://georust.org/)
+ecosystem. This support may be enabled using the `geo` feature flag. When the
+`geo` feature is enabled, some types will loose support for `serde`
+serialization & deserialization.
+
 * 2.2.0: 2022-09-03: This crate's `Waypoint` and `Location` types now have
 variants that represent the [geo](https://crates.io/crates/geo) crate's
 `Coordinate` and `Point` types. `Locations` type now has variants that represent
-rhe geo crate's `Line` and `LineString` types. It's now possible to make
+the geo crate's `Line` and `LineString` types. It's now possible to make
 most Google Maps API requests using geo types, using the provided special helper
 methods when the `geo` feature flag is enabled.
 
 * 2.2.0: 2022-08-27: Optional type conversion support for the
 [geo](https://crates.io/crates/geo) crate. This feature can be enabled with the
-`geo_types` feature flag. It makes using these crates together a little less
+`geo` feature flag. It makes using these crates together a little less
 burdensome. Includes some unidirectional and some bidirectional
 [TryFrom](https://doc.rust-lang.org/std/convert/trait.TryFrom.html) conversions
 between this crate's `LatLng`, `Waypoint`, `Bounds`, and geo's `Coordinate`,
@@ -29,7 +35,7 @@ take advantage of the new `phf` tables.
 
 * 2.1.7: 2022-08-27: Google Maps client types now implement `FromStr` which
 gives access to `parse`. For example:
-`let lat_lng: LatLng = "49.8845,-97.1468".parse()?;`
+`let the_golden_boy: LatLng = "49.8845224,-97.1469436".parse()?;`
 
 * 2.1.7: 2022-08-22: Added debug logging message to show Google Maps client's
 request activity.